Move creds logic to pkg
